refactor/extract setup to new module
was going to add test coverage, but stubbing:logic
ratio seems way off, stub lots to test like, an if
statement, so leaving untested for now…
Tested manually though.
